+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Version 1.7.0 (2025-04-xx)
 
 **Features**
-- Unified Game events timeline (sessions, dates) with event type selection
+- Unified Game events timeline (sessions, sets, dates) with sorting, event type selection
 - Insert single sets
 - Reload csv import filenames from last filename in import folder
 - Navigation bar restyle